refactores adding new event
